+++ app/javascript/components/countdown.js
@@ -1,0 +1,34 @@
+const initCountDown = () => {
+  const second = 1000,
+      minute = second * 60,
+      hour = minute * 60,
+      day = hour * 24;
+
+  // Sélectionner date et heure d'un tour
+  const tourDate = document.getElementById('tour-date');
+
+  if (tourDate) {
+    const setTourDate = tourDate.dataset.datetime;
+
+    let countDown = new Date(`${setTourDate}`).getTime(),
+    x = setInterval(function() {
+
+    let now = new Date().getTime(),
+    distance = countDown - now;
+
+    document.getElementById('days').innerText = Math.floor(distance / (day)),
+    document.getElementById('hours').innerText = Math.floor((distance % (day)) / (hour)),
+    document.getElementById('minutes').innerText = Math.floor((distance % (hour)) / (minute)),
+    document.getElementById('seconds').innerText = Math.floor((distance % (minute)) / second);
+
+    //do something later when date is reached
+    //if (distance < 0) {
+    //  clearInterval(x);
+    //  'IT'S MY BIRTHDAY!;
+    //}
+
+  }, second)
+};
+};
+
+export { initCountDown };
